added unit tests for small functions in utils file

+++ install.sh
@@ -8,24 +8,27 @@ path_loc="/usr/local/bin/"
 data_loc="/usr/lib/"
 
 meson setup $build_dir 
-meson compile -C $build_dir 
+if meson compile -C $build_dir; then;
 
 
-if ! test -w $path_loc; then 
-    rm -r $build_dir 
-    exit 1
+    if ! test -w $path_loc; then 
+        rm -r $build_dir 
+        exit 1
 
+    fi 
+
+    cp ${build_dir}/linuxTimeMachineBackend ${path_loc}timeMachineCLI
+    cp ${build_dir}/linuxTimeMachineApp ${path_loc}timeMachine
+
+    echo $PATH
+
+
+
+    mkdir -p ${data_loc}/TimeMachine/
+    cp ./data/*ui ${data_loc}/TimeMachine/
+
+else 
+    echo "Compilation failed"
 fi 
-
-cp ${build_dir}/linuxTimeMachineBackend ${path_loc}timeMachineCLI
-cp ${build_dir}/linuxTimeMachineApp ${path_loc}timeMachine
-
-echo $PATH
-
-
-
-mkdir -p ${data_loc}/TimeMachine/
-cp ./data/*ui ${data_loc}/TimeMachine/
-
 
 rm -r $build_dir
